services/hermes-runtime/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
openai>=2.8.0
google-genai==1.47.0
